requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
greenlet==2.0.2
idna==3.4
Naked==0.1.32
protobuf==3.20.3
pyasn1==0.5.0
pycryptodome==3.18.0
pydantic==1.10.7
PyYAML==6.0
requests==2.30.0
shellescape==3.8.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
tk==0.1.0
typing_extensions==4.5.0
urllib3==2.0.2
